Moved middleware to separate component/container.
