+++ src/request/containers/RequestDetailPanelRequestContainer.ts
@@ -1,4 +1,4 @@
-import { getMiddleware, getRequest, getRequestContentType, getResponseContentType } from '../selectors/RequestDetailRequestSelectors';
+import { getRequest, getRequestContentType, getResponseContentType } from '../selectors/RequestDetailRequestSelectors';
 import { Request } from '../components/RequestDetailPanelRequest';
 
 import { connect } from 'react-redux';
@@ -7,7 +7,6 @@ function mapStateToProps(state) {
     const request = getRequest(state);
 
     return {
-        middleware: getMiddleware(state),
         url: request.url,
         request: {
             body: request.request.body,
